% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENTREannotDb.R
\docType{class}
\name{CENTREannotDb}
\alias{CENTREannotDb}
\alias{CENTREannotDb-class}
\alias{show}
\alias{dbconn,CENTREannotDb-method}
\alias{show,CENTREannotDb-method}
\alias{CENTREannotationDb}
\alias{tables}
\title{Database for the CENTRE precomputed data}
\usage{
CENTREannotationDb(x)

tables(x)
}
\arguments{
\item{x}{sqlite file path}
}
\description{
The \code{CENTREannotDb} object provides access to the ENCODE screen V3 annotation
for Human https://screen.encodeproject.org.
}
\details{
\code{CENTREannotDb} is the object that provides access to CENTRE's database.
\code{CENTREannotDb@.properties$tables} shows the tables inside the database and
their columns.
}
