% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryEngine.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from the CENTREannotDb databases}
\usage{
fetch_data(x, columns, filter, column_filter)
}
\arguments{
\item{x}{\code{CENTREannotDb}}

\item{columns}{Columns to select. Equivalent to X in SELECT X.}

\item{filter}{Element ID to select. Equivalen to ID in SELECT X FROM TABLE
WHERE ID in Z. If filter or column_filter is missing the program assumes the
query is SELECT X FROM TABLE.}

\item{column_filter}{Column on which to apply filter. Equivalent to Z in
SELECT X FROM TABLE WHERE ID in Z. If filter or column_filter is missing the
program assumes the query is SELECT X FROM TABLE.}
}
\description{
Main interface to fetch data from the CENTREannotation package databases
through the \code{CENTREannotDb} objects.
}
\examples{
  res <- fetch_data(CENTREannotDb, 
                    columns=c("enhancer_id", "start"), 
                    filter=c("EH38E1519134","EH38E1519132") , 
                    column_filter="enhancer_id")
}
